AskPic-BE/requirements.txt
python-dotenv
google-generativeai
langchain
langchain-google-genai
fastapi
pillow